C++ Two Dimensional Array
Two Dimensional Array Program in C++
Two dimensional (2D) array can be made in C++ programming language by using two for loops, first is outer for loop and the second one is inner for loop. The outer for loop is responsible for rows and the inner for loop is responsible for columns as shown here in the following program.